Disable notification by default
The HipChat API has Notify defaulting to False for a reason. It causes unnecessary interruptions to users of HipChat rooms, and basically makes alerts useless as usually the alert is actionless information. This commit disables the the notification, reverting the behavior to the API default. If Notify is desired, it should be added as an option to be enabled, not hard-coded.